backend/requirements.txt
fastapi
uvicorn
celery
redis
pytest-cov
flake8
minio
python-dotenv
psycopg2-binary
sqlalchemy
alembic
databases
pydantic
asyncpg
pillow
python-multipart
httpx